requirements.txt
gradio
opencv-python
numpy
scikit-learn
matplotlib
Pillow
